ACPICA: DT compiler: prevent error if optional field at the end of table is not present
ACPICA commit a7f73af9003bf4e730db5a133300c01ba7992a01

The data table compiler throws under the following conditions:
1.) there is a table with a last field that is optional
2.) if the optional field is not present
3.) the optional field is the last line of the data table

A change was made to dt_compile_table to return an AE_EOF under these conditions.
This AE_EOF means that we are at the end of the file. The caller to dt_compile_table
is responsible for handling this case. For DBG2 table, we will complete the
compilation of this subtable. For other tables, this could be different.

Link: https://github.com/acpica/acpica/commit/a7f73af9
Signed-off-by: Erik Schmauss <erik.schmauss@intel.com>
Signed-off-by: Bob Moore <robert.moore@intel.com>